data/scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = "Fetch images + metadata from Unsplash"
authors = ["Larry D Almeida <hello@larrydalmeida.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
progress = "^1.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
